feat: Changes the database configuration to PostgreSQL and updates related dependencies

+++ migrations/001_initial.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS jobs (
     site_url TEXT NOT NULL,
     status TEXT NOT NULL DEFAULT 'pending',
     -- pending | crawling | analyzing | completed | failed
-    total_pages INTEGER NOT NULL DEFAULT 0,
-    processed_pages INTEGER NOT NULL DEFAULT 0,
+    total_pages BIGINT NOT NULL DEFAULT 0,
+    processed_pages BIGINT NOT NULL DEFAULT 0,
     created_at TEXT NOT NULL,
     updated_at TEXT NOT NULL,
     error_message TEXT
@@ -17,26 +17,26 @@ CREATE TABLE IF NOT EXISTS pages (
     job_id TEXT NOT NULL REFERENCES jobs(id) ON DELETE CASCADE,
     url TEXT NOT NULL,
     title TEXT,
-    word_count INTEGER NOT NULL DEFAULT 0,
+    word_count BIGINT NOT NULL DEFAULT 0,
     -- Scores GEO (0.0 – 1.0)
-    score_cite_sources REAL NOT NULL DEFAULT 0,
-    score_quotation_addition REAL NOT NULL DEFAULT 0,
-    score_statistics_addition REAL NOT NULL DEFAULT 0,
-    score_fluency REAL NOT NULL DEFAULT 0,
-    score_authoritative_tone REAL NOT NULL DEFAULT 0,
-    score_technical_terms REAL NOT NULL DEFAULT 0,
-    score_easy_to_understand REAL NOT NULL DEFAULT 0,
-    score_content_structure REAL NOT NULL DEFAULT 0,
-    score_metadata_quality REAL NOT NULL DEFAULT 0,
-    score_schema_markup REAL NOT NULL DEFAULT 0,
-    score_content_depth REAL NOT NULL DEFAULT 0,
-    geo_score REAL NOT NULL DEFAULT 0,
+    score_cite_sources DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_quotation_addition DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_statistics_addition DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_fluency DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_authoritative_tone DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_technical_terms DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_easy_to_understand DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_content_structure DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_metadata_quality DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_schema_markup DOUBLE PRECISION NOT NULL DEFAULT 0,
+    score_content_depth DOUBLE PRECISION NOT NULL DEFAULT 0,
+    geo_score DOUBLE PRECISION NOT NULL DEFAULT 0,
     -- Recomendações em JSON
     recommendations TEXT NOT NULL DEFAULT '[]',
     -- Metadados extras em JSON
     meta_description TEXT,
-    has_og_tags INTEGER NOT NULL DEFAULT 0,
-    has_schema_markup INTEGER NOT NULL DEFAULT 0,
+    has_og_tags BIGINT NOT NULL DEFAULT 0,
+    has_schema_markup BIGINT NOT NULL DEFAULT 0,
     analyzed_at TEXT NOT NULL
 );
