+++ migrations/002_llm_cache.sql
@@ -2,10 +2,10 @@
 -- Evita re-executar inferência para URLs já analisadas
 CREATE TABLE IF NOT EXISTS llm_cache (
     url                TEXT PRIMARY KEY,
-    fluency            REAL NOT NULL,
-    authoritative_tone REAL NOT NULL,
-    technical_terms    REAL NOT NULL,
-    easy_to_understand REAL NOT NULL,
+    fluency            DOUBLE PRECISION NOT NULL,
+    authoritative_tone DOUBLE PRECISION NOT NULL,
+    technical_terms    DOUBLE PRECISION NOT NULL,
+    easy_to_understand DOUBLE PRECISION NOT NULL,
     fluency_rec        TEXT,
     auth_rec           TEXT,
     tech_rec           TEXT,
